Huge refactor
