+++ helpers/logger.py
@@ -1,0 +1,20 @@
+import logging
+from logging.handlers import RotatingFileHandler
+
+def setup_logger(logger_name, log_file, level=logging.INFO):
+    logger = logging.getLogger(logger_name)
+    logger.setLevel(level)
+
+    handler = RotatingFileHandler(
+        log_file,
+        maxBytes=10 * 1024 * 1024,
+        backupCount=5,
+        encoding='utf-8',
+        mode='a'
+    )
+    formatter = logging.Formatter('%(asctime)s - %(name)s - %(levelname)s - %(message)s')
+    
+    handler.setFormatter(formatter)
+    logger.addHandler(handler)
+    
+    return logger
